requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==7.1.2
colorclass==2.2.2
colorlog==6.7.0
croniter==1.3.8
Django==4.1.7
django-filter==22.1
faust==1.10.4
frozenlist==1.3.3
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
graphqlclient==0.2.4
idna==3.4
kafka-python==1.4.7
mode==4.3.2
multidict==6.0.4
mypy-extensions==1.0.0
opentracing==1.3.0
promise==2.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
requests==2.28.2
robinhood-aiokafka==1.1.6
six==1.16.0
sqlparse==0.4.3
terminaltables==3.1.10
text-unidecode==1.3
urllib3==1.26.15
venusian==1.2.0
yarl==1.8.2
